+++ viadot/flows/sql_server_to_duckdb.py
@@ -2,7 +2,7 @@ from prefect import Flow
 from typing import Any, Dict, List, Literal
 
 
-from ..task_utils import df_to_parquet, add_ingestion_metadata_task
+from ..task_utils import df_to_parquet, add_ingestion_metadata_task, cast_df_to_str
 from ..tasks import SQLServerToDF, DuckDBCreateTableFromParquet
 
 df_task = SQLServerToDF()
@@ -18,6 +18,7 @@ class SQLServerToDuckDB(Flow):
         duckdb_table: str = None,
         duckdb_schema: str = None,
         if_exists: Literal["fail", "replace", "append", "skip", "delete"] = "fail",
+        if_empty: Literal["warn", "skip", "fail"] = "skip",
         duckdb_credentials: dict = None,
         *args: List[any],
         **kwargs: Dict[str, Any],
@@ -34,6 +35,7 @@ class SQLServerToDuckDB(Flow):
             duckdb_table (str, optional): Destination table in DuckDB. Defaults to None.
             duckdb_schema (str, optional): Destination schema in DuckDB. Defaults to None.
             if_exists (Literal, optional):  What to do if the table already exists. Defaults to "fail".
+            if_empty (Literal, optional): What to do if Parquet file is empty. Defaults to "skip".
             duckdb_credentials (dict, optional): Credentials for the DuckDB connection. Defaults to None.
 
         """
@@ -46,6 +48,7 @@ class SQLServerToDuckDB(Flow):
         self.duckdb_table = duckdb_table
         self.duckdb_schema = duckdb_schema
         self.if_exists = if_exists
+        self.if_empty = if_empty
         self.duckdb_credentials = duckdb_credentials
 
         super().__init__(*args, name=name, **kwargs)
@@ -61,9 +64,9 @@ class SQLServerToDuckDB(Flow):
             config_key=self.sqlserver_config_key, query=self.sql_query, flow=self
         )
         df_with_metadata = add_ingestion_metadata_task.bind(df, flow=self)
-
+        df_mapped = cast_df_to_str.bind(df_with_metadata, flow=self)
         parquet = df_to_parquet.bind(
-            df=df_with_metadata,
+            df=df_mapped,
             path=self.local_file_path,
             if_exists=self.if_exists,
             flow=self,
@@ -73,6 +76,7 @@ class SQLServerToDuckDB(Flow):
             schema=self.duckdb_schema,
             table=self.duckdb_table,
             if_exists=self.if_exists,
+            if_empty=self.if_empty,
             flow=self,
         )
         create_duckdb_table.set_upstream(parquet, flow=self)